requirements.txt
slack_bolt
python-dotenv
datadog-api-client